Seperating the middleware
